node_scripts/requirements.txt
azure-batch==3.0.0
azure-storage==0.33.0
pyyaml==3.12